+++ packages/gasket-preset-api/generator/server.js
@@ -1,6 +1,2 @@
 import gasket from './gasket.js';
-{{#if apiApp }}
-import './routes/index.js';
-{{/if}}
-
 gasket.actions.startServer();