+++ packages/gasket-plugin-fastify/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `@gasket/plugin-fastify`
 
+### 7.1.4
+
+- Deprecate and remove the use of the `getFastifyApp` action ([#1011])
+
 ### 7.1.0
 
 - Aligned version releases across all packages
@@ -88,3 +92,4 @@ Enable middleware support ([#172])
 [#940]: https://github.com/godaddy/gasket/pull/940
 [#969]: https://github.com/godaddy/gasket/pull/969
 [#972]: https://github.com/godaddy/gasket/pull/972
+[#1011]: https://github.com/godaddy/gasket/pull/1011