+++ packages/gasket-plugin-typescript/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `@gasket/plugin-typescript`
 
+### 7.1.4
+
+- Remove routes import from the `server.ts` ([#1011])
+
 ### 7.1.0
 
 - Add HTTPS proxy server to Next.js dev server ([#982])
@@ -13,3 +17,4 @@
 [Version 7 Upgrade Guide]: /docs/upgrade-to-7.md
 
 [#982]: https://github.com/godaddy/gasket/pull/982
+[#1011]: https://github.com/godaddy/gasket/pull/1011